R08/T02/feature/register_restaurant_page2 | Adição da appbar e título
- Chamei a classe RegisterPage no ícone do usuário na appbar da homepage temporariamente, apenas para visualizar melhor
- Adicionei uma appBar à classe
- Aicionei os títulos
